Implement the APIGW feedback mechanism
What
-------

This adds the implementation of sending feedback from the frontend for
each chat response. This shows users an example of how to make
Cognito-authenticated API requests from the frontend, through API Gateway,
and to a Lambda function.

How
----

This example also demonstrates a few critical things that are often
confusing or challenging:

- Using Lambda layers
- Integration, and use of, Lambda PowerTools, which implements many best
  practices
- CORS handling, via LPT and CDK updates (handles amplify hosting and
  localhost development)
- Payload validation, with Pydantic via LPT
- Authenticated API requests from the frontend.
- Add ShadCN dialog component